+++ tests/cli/deploy_test.go
@@ -1,6 +1,9 @@
 package cli_test
 
-import "testing"
+import (
+	"errors"
+	"testing"
+)
 
 func TestDeploy(t *testing.T) {
 	run(t, []Command{
@@ -72,6 +75,37 @@ Status: Finished processing events for release v1 of acme-inc`,
 		{
 			"releases -a acme-inc",
 			"v1    Dec 31  2014  Deploy remind101/acme-inc:latest (fake)",
+		},
+	})
+}
+
+func TestDeploy_CommitMessageRequired(t *testing.T) {
+	pre := func(cli *CLI) {
+		cli.Empire.MessagesRequired = true
+	}
+
+	runWithPre(t, pre, []Command{
+		{
+			"deploy remind101/acme-inc",
+			errors.New("error: A message is required for this action, please run again with '-m'."),
+		},
+	})
+
+	runWithPre(t, pre, []Command{
+		{
+			"deploy remind101/acme-inc -m commit",
+			`Pulling repository remind101/acme-inc
+345c7524bc96: Pulling image (latest) from remind101/acme-inc
+345c7524bc96: Pulling image (latest) from remind101/acme-inc, endpoint: https://registry-1.docker.io/v1/
+345c7524bc96: Pulling dependent layers
+a1dd7097a8e8: Download complete
+Status: Image is up to date for remind101/acme-inc:latest
+Status: Created new release v1 for acme-inc
+Status: Finished processing events for release v1 of acme-inc`,
+		},
+		{
+			"releases -a acme-inc",
+			"v1    Dec 31  2014  Deploy remind101/acme-inc:latest (fake: 'commit')",
 		},
 	})
 }